Backend/enhc: linking vehicle to driver issues and refactoring Status.hs to remove the types and vehicle status updates to new file
