+++ Backend/app/provider-platform/dynamic-offer-driver-app/Main/dynamic-offer-driver-app.cabal
@@ -498,6 +498,7 @@ library
       SharedLogic.DriverOnboarding
       SharedLogic.DriverOnboarding.Digilocker
       SharedLogic.DriverOnboarding.Status
+      SharedLogic.DriverOnboarding.VehicleDocs
       SharedLogic.DriverPool
       SharedLogic.DriverPool.Config
       SharedLogic.DriverPool.DriverPoolData
