+++ Backend/app/provider-platform/dynamic-offer-driver-app/Main/src/SharedLogic/DriverOnboarding/Status.hs
@@ -13,6 +13,7 @@ module SharedLogic.DriverOnboarding.Status
     mapStatus,
     checkAllVehicleDocsVerifiedForRC,
     checkAllDriverDocsVerifiedForDriver,
+    validateMandatoryVehicleDocsForRC,
     activateRCAutomatically,
     mkCommonDocumentItem,
     checkInspectionHubRequestCreated,
@@ -36,7 +37,6 @@ import Data.List (nub)
 import qualified Data.Text as T
 import qualified Domain.Action.UI.DriverOnboarding.DriverLicense as DDL
 import qualified Domain.Action.UI.DriverOnboarding.VehicleRegistrationCertificate as DomainRC
-import qualified Domain.Action.UI.Plan as DAPlan
 import qualified Domain.Types.AadhaarCard as DAadhaarCard
 import qualified Domain.Types.CommonDriverOnboardingDocuments as DCDOD
 import qualified Domain.Types.DocStatus as DocStatus
@@ -45,19 +45,14 @@ import qualified Domain.Types.DocumentVerificationConfig as DDVC
 import qualified Domain.Types.DocumentVerificationConfig as DVC
 import qualified Domain.Types.DriverInformation as DI
 import qualified Domain.Types.DriverLicense as DL
-import qualified Domain.Types.FleetOwnerDocumentVerificationConfig as FODVC
 import qualified Domain.Types.FleetOwnerInformation as DFOI
-import qualified Domain.Types.HyperVergeVerification as HV
-import qualified Domain.Types.IdfyVerification as IV
 import qualified Domain.Types.Merchant as DM
 import qualified Domain.Types.MerchantOperatingCity as DMOC
 import qualified Domain.Types.OperationHubRequests as DOHR
 import qualified Domain.Types.Person as DP
-import Domain.Types.Plan as Plan
 import qualified Domain.Types.TransporterConfig as DTC
 import qualified Domain.Types.VehicleCategory as DVC
 import qualified Domain.Types.VehicleRegistrationCertificate as RC
-import qualified Domain.Types.VehicleVariant as DV
 import Environment
 import GHC.Records.Extra (HasField)
 import Kernel.Beam.Functions
@@ -73,6 +68,7 @@ import Kernel.Types.Id
 import Kernel.Utils.Common hiding (HasField)
 import qualified SharedLogic.DriverOnboarding as SDO
 import qualified SharedLogic.DriverOnboarding.Digilocker as SDDigilocker
+import SharedLogic.DriverOnboarding.VehicleDocs
 import qualified SharedLogic.MessageBuilder as MessageBuilder
 import qualified Storage.Beam.IssueManagement ()
 import qualified Storage.Cac.TransporterConfig as SCTC
@@ -97,24 +93,15 @@ import qualified Storage.Queries.FleetOwnerInformation as QFOI
 import qualified Storage.Queries.HyperVergeVerification as HVQuery
 import qualified Storage.Queries.IdfyVerification as IVQuery
 import qualified Storage.Queries.Image as IQuery
-import qualified Storage.Queries.OperationHubRequestsExtra as QOHRE
 import qualified Storage.Queries.Person as QPerson
 import qualified Storage.Queries.RideExtra as QRideExtra
-import qualified Storage.Queries.Translations as MTQuery
 import qualified Storage.Queries.Vehicle as QVehicle
-import qualified Storage.Queries.VehicleFitnessCertificate as VFCQuery
-import qualified Storage.Queries.VehicleInsurance as VIQuery
-import qualified Storage.Queries.VehicleNOC as VNOCQuery
-import qualified Storage.Queries.VehiclePUC as VPUCQuery
-import qualified Storage.Queries.VehiclePermit as VPQuery
 import qualified Storage.Queries.VehicleRegistrationCertificate as RCQuery
 import qualified Tools.BackgroundVerification as BackgroundVerification
-import Tools.Error (DriverOnboardingError (ImageNotValid))
 import qualified Tools.Plasma as TPlasma
 import qualified Tools.SMS as Sms
 import qualified Tools.Verification as Verification
 
-type DocVerificationConfigs = Either [FODVC.FleetOwnerDocumentVerificationConfig] [DVC.DocumentVerificationConfig]
 
 data PersonStatusContext = PersonStatusContext
   { statusPerson :: DP.Person,
@@ -143,8 +130,6 @@ data StatusEvent
 --   which the doc switches to when, for example, it's expired or when it is invalidated from dashboard.
 -- PULL_REQUIRED is used when a document needs to be pulled from DigiLocker
 -- CONSENT_DENIED is used when user denies consent for DigiLocker verification
-data ResponseStatus = NO_DOC_AVAILABLE | PENDING | VALID | FAILED | INVALID | LIMIT_EXCEED | MANUAL_VERIFICATION_REQUIRED | UNAUTHORIZED | PULL_REQUIRED | CONSENT_DENIED
-  deriving (Show, Eq, Generic, ToJSON, FromJSON, ToSchema, ToParamSchema, Enum, Bounded)
 
 data StatusRes' = StatusRes'
   { driverDocuments :: [DocumentStatusItem],
@@ -157,34 +142,7 @@ data StatusRes' = StatusRes'
     digilockerAuthorizationUrl :: Maybe Text
   }
 
-data VehicleDocumentItem = VehicleDocumentItem
-  { registrationNo :: Text,
-    userSelectedVehicleCategory :: DVC.VehicleCategory,
-    verifiedVehicleCategory :: Maybe DVC.VehicleCategory,
-    isVerified :: Bool,
-    isActive :: Bool,
-    isApproved :: Bool,
-    vehicleModel :: Maybe Text,
-    documents :: [DocumentStatusItem],
-    dateOfUpload :: UTCTime,
-    s3Path :: Maybe Text,
-    imageId :: Maybe Text,
-    documentExpiry :: Maybe UTCTime,
-    docsVerificationStatus :: Maybe DDVS.DocsVerificationStatus
-  }
-  deriving (Show, Eq, Generic, ToJSON, FromJSON, ToSchema)
 
-data DocumentStatusItem = DocumentStatusItem
-  { documentType :: DDVC.DocumentType,
-    verificationStatus :: ResponseStatus,
-    verificationMessage :: Maybe Text,
-    verificationUrl :: Maybe BaseUrl,
-    s3Path :: Maybe Text,
-    imageId :: Maybe Text,
-    imageId2 :: Maybe Text,
-    documentExpiry :: Maybe UTCTime
-  }
-  deriving (Show, Eq, Generic, ToJSON, FromJSON, ToSchema)
 
 data CommonDocumentItem = CommonDocumentItem
   { documentType :: DDVC.DocumentType,
@@ -563,7 +521,11 @@ statusHandler' person entityImagesInfo makeSelfieAadhaarPanMandatory prefillData
               getVehicleDocTypes merchantOpCityId driverDocConfigs v.verifiedVehicleCategory v.userSelectedVehicleCategory (Just True)
             pure $ v {documents = filter (\d -> d.documentType `elem` mandatoryTypesForVehicle) v.documents}
           pure (driverDocumentsMandatory, vehicleDocumentsMandatory)
-    persistDocsVerificationStatuses person driverDocsForPersist vehicleDocsForPersist
+    let vehicleDocsForPersist' =
+          if transporterConfig.separateDriverVehicleEnablement == Just True && person.role == DP.DRIVER
+            then []
+            else vehicleDocsForPersist
+    persistDocsVerificationStatuses person driverDocsForPersist vehicleDocsForPersist'
 
   enabled <-
     if isFleetRole person.role
@@ -713,78 +675,7 @@ fetchDriverDocuments entityImagesInfo allDocVerificationConfigs possibleVehicleC
     let finalMessage = mbReason <|> (if isDigiLockerEnabled then responseCode else Nothing) <|> mbMessage
     return $ DocumentStatusItem {documentType = docType, verificationStatus = status, verificationMessage = finalMessage, verificationUrl = mbUrl, s3Path = mbS3PathFinal, imageId = mbImageIdFinal, imageId2 = mbImageId2Final, documentExpiry = mbExpiryFinal}
 
-fetchVehicleDocuments ::
-  IQuery.EntityImagesInfo ->
-  [DVC.DocumentVerificationConfig] ->
-  Language ->
-  Maybe Text ->
-  Maybe Bool ->
-  Bool ->
-  Flow [VehicleDocumentItem]
-fetchVehicleDocuments entityImagesInfo allDocumentVerificationConfigs language Nothing onlyMandatoryDocs skipMessages = do
-  -- All items required
-  processedVehicleDocumentsWithRC <- fetchProcessedVehicleDocumentsWithRC entityImagesInfo allDocumentVerificationConfigs language Nothing onlyMandatoryDocs skipMessages
-  processedVehicleDocumentsWithoutRC <- fetchProcessedVehicleDocumentsWithoutRC entityImagesInfo allDocumentVerificationConfigs processedVehicleDocumentsWithRC Nothing onlyMandatoryDocs
-  let processedVehicleDocuments = processedVehicleDocumentsWithoutRC <> processedVehicleDocumentsWithRC
-  inprogressVehicleDocuments <- fetchInprogressVehicleDocuments entityImagesInfo allDocumentVerificationConfigs language processedVehicleDocuments Nothing onlyMandatoryDocs skipMessages
-  pure $ processedVehicleDocuments <> inprogressVehicleDocuments
-fetchVehicleDocuments entityImagesInfo allDocumentVerificationConfigs language (Just reqRegistrationNo) onlyMandatoryDocs skipMessages = do
-  -- Only one item required with specific registrationNo
-  processedVehicleDocumentsWithRC <- fetchProcessedVehicleDocumentsWithRC entityImagesInfo allDocumentVerificationConfigs language (Just reqRegistrationNo) onlyMandatoryDocs skipMessages
-  let (entityTxt, entityId) = case entityImagesInfo.entity of
-        IQuery.PersonEntity driver -> ("driverId", driver.id.getId)
-        IQuery.VehicleRCEntity rc -> ("rcId", rc.id.getId)
-  if null processedVehicleDocumentsWithRC
-    then do
-      processedVehicleDocumentsWithoutRC <- fetchProcessedVehicleDocumentsWithoutRC entityImagesInfo allDocumentVerificationConfigs processedVehicleDocumentsWithRC (Just reqRegistrationNo) onlyMandatoryDocs
-      if null processedVehicleDocumentsWithoutRC
-        then do
-          docs <- fetchInprogressVehicleDocuments entityImagesInfo allDocumentVerificationConfigs language [] (Just reqRegistrationNo) onlyMandatoryDocs skipMessages
-          if null docs
-            then logWarning $ "No docs found for rcNo and " <> entityTxt <> ": " <> entityId
-            else logInfo $ "Inprogress vehicle docs found for rcNo and " <> entityTxt <> ": " <> entityId
-          pure docs
-        else do
-          logInfo $ "Processed vehicle docs without RC found for rcNo and " <> entityTxt <> ": " <> entityId
-          pure processedVehicleDocumentsWithoutRC
-    else do
-      logInfo $ "Processed vehicle documents with RC found for rcNo and " <> entityTxt <> ": " <> entityId
-      pure processedVehicleDocumentsWithRC
 
-getVehicleDocTypes ::
-  (Monad m, Log m) =>
-  Id DMOC.MerchantOperatingCity ->
-  [DVC.DocumentVerificationConfig] ->
-  Maybe DVC.VehicleCategory ->
-  DVC.VehicleCategory ->
-  Maybe Bool ->
-  m [DVC.DocumentType]
-getVehicleDocTypes merchantOpCityId allDocumentVerificationConfigs verifiedVehicleCategory userSelectedVehicleCategory onlyMandatoryDocs = do
-  let vehicleCategory = fromMaybe userSelectedVehicleCategory verifiedVehicleCategory
-      isVehicleSideDoc config = case config.documentCategory of
-        Just DDVC.Vehicle -> True
-        Just _ -> False
-        Nothing -> config.documentType `elem` SDO.defaultVehicleDocumentTypes
-      mandatoryVehicleDocumentVerificationConfigs =
-        filter
-          ( \config ->
-              fromMaybe config.isMandatory config.isMandatoryForEnabling
-                && config.vehicleCategory == vehicleCategory
-                && isVehicleSideDoc config
-          )
-          allDocumentVerificationConfigs
-  if onlyMandatoryDocs == Just True
-    then do
-      let vehicleDocumentTypes = nub (mandatoryVehicleDocumentVerificationConfigs <&> (.documentType))
-      logInfo $
-        "Fetch only mandatory vehicle docs types: merchantOpCityId: "
-          <> merchantOpCityId.getId
-          <> "; vehicleCategory: "
-          <> show vehicleCategory
-          <> "; vehicleDocumentTypes: "
-          <> show vehicleDocumentTypes
-      pure vehicleDocumentTypes
-    else pure SDO.defaultVehicleDocumentTypes
 
 getDriverDocTypes ::
   Id DMOC.MerchantOperatingCity ->
@@ -846,202 +737,8 @@ getDriverDocTypes merchantOpCityId allDocVerificationConfigs possibleVehicleCate
           pure driverDocumentTypes
         else pure SDO.defaultDriverDocumentTypes
 
-fetchProcessedVehicleDocumentsWithRC ::
-  IQuery.EntityImagesInfo ->
-  [DVC.DocumentVerificationConfig] ->
-  Language ->
-  Maybe Text ->
-  Maybe Bool ->
-  Bool ->
-  Flow [VehicleDocumentItem]
-fetchProcessedVehicleDocumentsWithRC entityImagesInfo allDocumentVerificationConfigs language mbReqRegistrationNo onlyMandatoryDocs skipMessages = do
-  let merchantOpCityId = entityImagesInfo.merchantOperatingCity.id
-  processedVehicles <- case entityImagesInfo.entity of
-    IQuery.PersonEntity person -> do
-      associations <- DRAQuery.findAllLinkedByDriverId person.id
-      (catMaybes <$>) $
-        forM associations $ \assoc -> do
-          mbRc <- RCQuery.findById assoc.rcId
-          -- filter by rcNo if required
-          mbFilteredRc <- case mbRc of
-            Just rc -> do
-              rcCertificateNumber <- decrypt rc.certificateNumber
-              let wrongRcNo = isJust mbReqRegistrationNo && Just rcCertificateNumber /= mbReqRegistrationNo
-              return $ if wrongRcNo then Nothing else Just rc
-            Nothing -> return Nothing
-          return $ (assoc.isRcActive,assoc.rcId,) <$> mbFilteredRc
-    IQuery.VehicleRCEntity rc -> do
-      mbAssoc <- DRAQuery.findLatestLinkedByRCId rc.id entityImagesInfo.now
-      pure [(maybe False (.isRcActive) mbAssoc, rc.id, rc)]
-  processedVehicles `forM` \(isActive, rcId, processedVehicle) -> do
-    rcImagesInfo <- IQuery.getRcImagesInfoFromEntityImagesInfo entityImagesInfo rcId vehicleDocsLoadedByRcId
-    registrationNo <- decrypt processedVehicle.certificateNumber
-    let dateOfUpload = processedVehicle.createdAt
-    let verifiedVehicleCategory = DV.castVehicleVariantToVehicleCategory <$> processedVehicle.vehicleVariant
-        userSelectedVehicleCategory = fromMaybe DVC.CAR $ processedVehicle.userPassedVehicleCategory <|> verifiedVehicleCategory
-        docVerificationConfigs = filter (\config -> config.vehicleCategory == userSelectedVehicleCategory) allDocumentVerificationConfigs
 
-    vehicleDocumentTypes <- getVehicleDocTypes merchantOpCityId allDocumentVerificationConfigs verifiedVehicleCategory userSelectedVehicleCategory onlyMandatoryDocs
-    documents <-
-      vehicleDocumentTypes `forM` \docType -> do
-        (mbStatus, mbProcessedReason, mbProcessedUrl, mbExpiry, mbS3Path, mbImageId) <- getProcessedVehicleDocuments entityImagesInfo docType processedVehicle (Just rcImagesInfo)
-        case mbStatus of
-          Just status -> do
-            mbMessage <- documentStatusMessage status Nothing docType mbProcessedUrl language skipMessages
-            return $ DocumentStatusItem {documentType = docType, verificationStatus = status, verificationMessage = mbProcessedReason <|> mbMessage, verificationUrl = mbProcessedUrl, s3Path = mbS3Path, imageId = mbImageId, imageId2 = Nothing, documentExpiry = mbExpiry}
-          Nothing -> do
-            (status, mbReason, mbUrl, _, mbS3PathInProgress, mbImageIdInProgress) <- getInProgressVehicleDocuments entityImagesInfo (Just rcImagesInfo) docType docVerificationConfigs
-            mbMessage <- documentStatusMessage status mbReason docType mbUrl language skipMessages
-            return $ DocumentStatusItem {documentType = docType, verificationStatus = status, verificationMessage = mbMessage, verificationUrl = mbUrl, s3Path = mbS3PathInProgress, imageId = mbImageIdInProgress, imageId2 = Nothing, documentExpiry = mbExpiry}
 
-    let mbRcImage = find (\img -> img.id == processedVehicle.documentImageId) entityImagesInfo.entityImages
-        rcS3Path = mbRcImage <&> (.s3Path)
-        rcImageId = Just processedVehicle.documentImageId.getId
-        rcExpiry = Just processedVehicle.fitnessExpiry
-    return
-      VehicleDocumentItem
-        { registrationNo,
-          userSelectedVehicleCategory,
-          verifiedVehicleCategory,
-          isVerified = False,
-          isActive,
-          isApproved = fromMaybe False processedVehicle.approved,
-          vehicleModel = processedVehicle.vehicleModel,
-          documents,
-          dateOfUpload,
-          s3Path = rcS3Path,
-          imageId = rcImageId,
-          documentExpiry = rcExpiry,
-          docsVerificationStatus = Just $ computeAdminDocsVerificationStatus documents
-        }
-
-fetchProcessedVehicleDocumentsWithoutRC ::
-  IQuery.EntityImagesInfo ->
-  [DVC.DocumentVerificationConfig] ->
-  [VehicleDocumentItem] ->
-  Maybe Text ->
-  Maybe Bool ->
-  Flow [VehicleDocumentItem]
-fetchProcessedVehicleDocumentsWithoutRC entityImagesInfo allDocumentVerificationConfigs processedVehicleDocumentsWithRC mbReqRegistrationNo onlyMandatoryDocs = do
-  let merchantOpCityId = entityImagesInfo.merchantOperatingCity.id
-      mbPersonId = IQuery.getPersonEntityId entityImagesInfo
-  mbVehicle <- maybe (pure Nothing) QVehicle.findById mbPersonId
-  case mbVehicle of
-    Just vehicle -> do
-      let vehicleAlreadyIncluded = isJust $ find (\doc -> doc.registrationNo == vehicle.registrationNo) processedVehicleDocumentsWithRC
-      -- filter by rcNo if required
-      let wrongRcNo = isJust mbReqRegistrationNo && Just vehicle.registrationNo /= mbReqRegistrationNo
-      if vehicleAlreadyIncluded || wrongRcNo
-        then return []
-        else do
-          let userSelectedVehicleCategory = DV.castVehicleVariantToVehicleCategory vehicle.variant
-              verifiedVehicleCategory = Just $ DV.castVehicleVariantToVehicleCategory vehicle.variant
-          vehicleDocumentTypes <- getVehicleDocTypes merchantOpCityId allDocumentVerificationConfigs verifiedVehicleCategory userSelectedVehicleCategory onlyMandatoryDocs
-
-          documents <-
-            vehicleDocumentTypes `forM` \docType -> do
-              return $ DocumentStatusItem {documentType = docType, verificationStatus = NO_DOC_AVAILABLE, verificationMessage = Nothing, verificationUrl = Nothing, s3Path = Nothing, imageId = Nothing, imageId2 = Nothing, documentExpiry = Nothing}
-          return
-            [ VehicleDocumentItem
-                { registrationNo = vehicle.registrationNo,
-                  userSelectedVehicleCategory,
-                  verifiedVehicleCategory,
-                  isVerified = True,
-                  isActive = True,
-                  isApproved = False,
-                  vehicleModel = Just vehicle.model,
-                  documents,
-                  dateOfUpload = vehicle.createdAt,
-                  s3Path = Nothing,
-                  imageId = Nothing,
-                  documentExpiry = Nothing,
-                  docsVerificationStatus = Nothing
-                }
-            ]
-    Nothing -> return []
-
-fetchInprogressVehicleDocuments ::
-  IQuery.EntityImagesInfo ->
-  [DVC.DocumentVerificationConfig] ->
-  Language ->
-  [VehicleDocumentItem] ->
-  Maybe Text ->
-  Maybe Bool ->
-  Bool ->
-  Flow [VehicleDocumentItem]
-fetchInprogressVehicleDocuments entityImagesInfo allDocumentVerificationConfigs language processedVehicleDocuments mbReqRegistrationNo onlyMandatoryDocs skipMessages = do
-  let merchantOpCityId = entityImagesInfo.merchantOperatingCity.id
-      mbPersonId = IQuery.getPersonEntityId entityImagesInfo
-  mbVerificationReqRecord <- case entityImagesInfo.entity of
-    IQuery.PersonEntity person -> do
-      -- Driver inspection or status handler
-      inprogressVehicleIdfy <- listToMaybe <$> IVQuery.findLatestByDriverIdAndDocType (Just 1) Nothing person.id DVC.VehicleRegistrationCertificate
-      inprogressVehicleHV <- listToMaybe <$> HVQuery.findLatestByDriverIdAndDocType (Just 1) Nothing person.id DVC.VehicleRegistrationCertificate
-      pure $ getLatestVerificationRecord inprogressVehicleIdfy inprogressVehicleHV
-    IQuery.VehicleRCEntity _rc -> do
-      -- Vehicle inspection
-      -- Inprogress rc doc for Vehicle inspection case not required, because we already have processed rc
-      pure Nothing
-  case mbVerificationReqRecord of
-    Just verificationReqRecord -> do
-      registrationNoEither <- withTryCatch "decryptDocumentNumber:fetchInprogressVehicleDocuments" (decrypt verificationReqRecord.documentNumber)
-      case registrationNoEither of
-        Left err -> do
-          logError $ "Error while decrypting document number: " <> (verificationReqRecord.documentNumber & unEncrypted . encrypted) <> " with err: " <> show err
-          return []
-        Right registrationNo -> do
-          -- filter by rcNo if required
-          let wrongRcNo = isJust mbReqRegistrationNo && Just registrationNo /= mbReqRegistrationNo
-          if wrongRcNo
-            then return []
-            else do
-              rcNoEnc <- encrypt registrationNo
-              rc <- RCQuery.findByCertificateNumberHash (rcNoEnc & hash)
-              (isUnlinked, mbRcId) <- case rc of
-                Just rc_ -> do
-                  iu <- maybe (pure []) (\personId -> DRAQuery.findUnlinkedRC personId rc_.id) mbPersonId
-                  pure (iu, Just rc_.id)
-                Nothing -> pure ([], Nothing)
-              mbRcImagesInfo <- forM mbRcId $ \rcId -> do
-                IQuery.getRcImagesInfoFromEntityImagesInfo entityImagesInfo rcId vehicleDocsLoadedByRcId
-              if isJust (find (\doc -> doc.registrationNo == registrationNo) processedVehicleDocuments) || not (null isUnlinked)
-                then return []
-                else do
-                  let userSelectedVehicleCategory = fromMaybe DVC.CAR verificationReqRecord.vehicleCategory
-                      verifiedVehicleCategory = Nothing
-                      docVerificationConfigs = filter (\config -> config.vehicleCategory == userSelectedVehicleCategory) allDocumentVerificationConfigs
-
-                  vehicleDocumentTypes <- getVehicleDocTypes merchantOpCityId allDocumentVerificationConfigs verifiedVehicleCategory userSelectedVehicleCategory onlyMandatoryDocs
-                  documents <-
-                    vehicleDocumentTypes `forM` \docType -> do
-                      (status, mbReason, mbUrl, _, mbS3Path, mbImageId) <- getInProgressVehicleDocuments entityImagesInfo mbRcImagesInfo docType docVerificationConfigs
-                      mbMessage <- documentStatusMessage status mbReason docType mbUrl language skipMessages
-                      return $ DocumentStatusItem {documentType = docType, verificationStatus = status, verificationMessage = mbMessage, verificationUrl = mbUrl, s3Path = mbS3Path, imageId = mbImageId, imageId2 = Nothing, documentExpiry = Nothing}
-                  let mbRcIdText = (.getId) <$> mbRcId
-                      mbRcImage =
-                        find
-                          (\img -> img.imageType == DVC.VehicleRegistrationCertificate && (isNothing mbRcIdText || img.rcId == mbRcIdText))
-                          entityImagesInfo.entityImages
-                      rcS3Path = mbRcImage <&> (.s3Path)
-                      rcImageId = (.id.getId) <$> mbRcImage
-                  return
-                    [ VehicleDocumentItem
-                        { registrationNo,
-                          userSelectedVehicleCategory,
-                          verifiedVehicleCategory,
-                          isVerified = False,
-                          isActive = False,
-                          isApproved = False,
-                          vehicleModel = Nothing,
-                          documents,
-                          dateOfUpload = verificationReqRecord.createdAt,
-                          s3Path = rcS3Path,
-                          imageId = rcImageId,
-                          documentExpiry = Nothing,
-                          docsVerificationStatus = Nothing
-                        }
-                    ]
-    Nothing -> return []
 
 checkAllVehicleDocsVerified ::
   [DVC.DocumentVerificationConfig] ->
@@ -1122,12 +819,6 @@ sendEnablementSms merchantOpCityId personId transporterConfig merchantId =
           MessageBuilder.BuildOnboardingMessageReq {}
       Sms.sendSMS merchantId merchantOpCityId (Sms.SendSMSReq message phoneNumber (fromMaybe sender mbSender) templateId messageType) >>= Sms.checkSmsResult
 
-computeAdminDocsVerificationStatus :: [DocumentStatusItem] -> DDVS.DocsVerificationStatus
-computeAdminDocsVerificationStatus docs
-  | null docs = DDVS.ADMIN_PENDING
-  | any ((`elem` [INVALID, UNAUTHORIZED]) . (.verificationStatus)) docs = DDVS.ADMIN_REJECTED
-  | all ((== VALID) . (.verificationStatus)) docs = DDVS.ADMIN_APPROVED
-  | otherwise = DDVS.ADMIN_PENDING
 
 persistDocsVerificationStatuses ::
   DP.Person ->
@@ -1348,85 +1039,8 @@ callGetDLGetStatus driverId merchantOpCityId = do
           unless ("still being processed" `T.isInfixOf` (fromMaybe "" resp.message)) (void $ DDL.onVerifyDL (SDO.makeHVVerificationReqRecord verificationReq) resp KEV.HyperVergeRCDL)
         _ -> throwError $ InternalError "Document and apiEndpoint mismatch occurred !!!!!!!!"
 
-getProcessedVehicleDocuments :: IQuery.EntityImagesInfo -> DVC.DocumentType -> RC.VehicleRegistrationCertificate -> Maybe IQuery.RcImagesInfo -> Flow (Maybe ResponseStatus, Maybe Text, Maybe BaseUrl, Maybe UTCTime, Maybe Text, Maybe Text)
-getProcessedVehicleDocuments entityImagesInfo docType vehicleRC mbRcImagesInfo = do
-  let entity = entityImagesInfo.entity
-      (mbS3Path, mbImageId) = getImageMetaFromVehicleImage entityImagesInfo docType mbRcImagesInfo
-      lookupImage imgId =
-        let mbImg = find (\img -> img.id == imgId) entityImagesInfo.entityImages
-            s3 = mbImg <&> (.s3Path)
-            iid = Just imgId.getId
-         in (s3, iid)
-  case docType of
-    DVC.VehicleRegistrationCertificate -> do
-      let status = mapStatus vehicleRC.verificationStatus
-          reason
-            | status == INVALID && not (null vehicleRC.failedRules) = Just $ T.intercalate ", " vehicleRC.failedRules
-            | otherwise = vehicleRC.rejectReason
-          (s3, iid) = lookupImage vehicleRC.documentImageId
-      return (Just status, reason, Nothing, Just vehicleRC.fitnessExpiry, s3, iid)
-    DVC.VehiclePermit -> do
-      mbDoc <- listToMaybe <$> VPQuery.findByRcId (Just 1) Nothing vehicleRC.id
-      let (s3, iid) = maybe (mbS3Path, mbImageId) (lookupImage . (.documentImageId)) mbDoc
-      return (mapStatus <$> (mbDoc <&> (.verificationStatus)), Nothing, Nothing, vehicleRC.permitExpiry, s3, iid)
-    DVC.VehicleFitnessCertificate -> do
-      mbDoc <- listToMaybe <$> VFCQuery.findByRcId (Just 1) Nothing vehicleRC.id
-      let (s3, iid) = maybe (mbS3Path, mbImageId) (lookupImage . (.documentImageId)) mbDoc
-      return (mapStatus <$> (mbDoc <&> (.verificationStatus)), Nothing, Nothing, Just vehicleRC.fitnessExpiry, s3, iid)
-    DVC.VehicleInsurance -> do
-      mbDoc <- listToMaybe <$> VIQuery.findByRcId (Just 1) Nothing vehicleRC.id
-      let (s3, iid) = maybe (mbS3Path, mbImageId) (lookupImage . (.documentImageId)) mbDoc
-      return (mapStatus <$> (mbDoc <&> (.verificationStatus)), (mbDoc >>= (.rejectReason)), Nothing, vehicleRC.insuranceValidity, s3, iid)
-    DVC.VehiclePUC -> do
-      mbDoc <- listToMaybe <$> VPUCQuery.findByRcId (Just 1) Nothing vehicleRC.id
-      let (s3, iid) = maybe (mbS3Path, mbImageId) (lookupImage . (.documentImageId)) mbDoc
-      return (mapStatus <$> (mbDoc <&> (.verificationStatus)), Nothing, Nothing, vehicleRC.pucExpiry, s3, iid)
-    DVC.VehicleNOC -> do
-      mbDoc <- listToMaybe <$> VNOCQuery.findByRcId (Just 1) Nothing vehicleRC.id
-      let (s3, iid) = maybe (mbS3Path, mbImageId) (lookupImage . (.documentImageId)) mbDoc
-      return (mapStatus <$> (mbDoc <&> (.verificationStatus)), Nothing, Nothing, mbDoc <&> (.nocExpiry), s3, iid)
-    DVC.VehicleInspectionForm -> do
-      -- Check all vehicle photos based on RC, not driver
-      (status, reason, url) <- checkVehiclePhotosStatusByRC mbRcImagesInfo
-      return (Just status, reason, url, Nothing, Nothing, Nothing)
-    DVC.InspectionHub -> do
-      registrationNo <- decrypt vehicleRC.certificateNumber
-      status <- getInspectionHubStatusForResponseStatus DOHR.ONBOARDING_INSPECTION Nothing (Just registrationNo)
-      return (status, Nothing, Nothing, Nothing, Nothing, Nothing)
-    DVC.SubscriptionPlan -> case entity of
-      IQuery.PersonEntity person -> do
-        mbPlan <- snd <$> DAPlan.getSubcriptionStatusWithPlan Plan.YATRI_SUBSCRIPTION person.id -- fix later on basis of vehicle category
-        return (Just $ boolToStatus (isJust mbPlan), Nothing, Nothing, Nothing, Nothing, Nothing)
-      IQuery.VehicleRCEntity _rc -> return (Nothing, Nothing, Nothing, Nothing, Nothing, Nothing)
-    _ -> do
-      let mbLatestImage = case mbRcImagesInfo of
-            Just rcImagesInfo -> listToMaybe (IQuery.filterRecentByPersonRCAndImageType rcImagesInfo docType)
-            Nothing -> Nothing
-          mbStatus = mbLatestImage >>= (.verificationStatus) <&> mapStatus
-      return (mbStatus, Nothing, Nothing, Nothing, mbS3Path, mbImageId)
-  where
-    boolToStatus :: Bool -> ResponseStatus
-    boolToStatus = \case
-      True -> VALID
-      False -> NO_DOC_AVAILABLE
 
-getImageMetaFromLatestImage :: IQuery.EntityImagesInfo -> DVC.DocumentType -> (Maybe Text, Maybe Text)
-getImageMetaFromLatestImage entityImagesInfo docType =
-  let images = IQuery.filterRecentByEntityIdAndImageType entityImagesInfo docType
-      mbLatestImage = listToMaybe images
-   in (mbLatestImage <&> (.s3Path), mbLatestImage <&> (.id.getId))
 
-getImageMetaFromVehicleImage :: IQuery.EntityImagesInfo -> DVC.DocumentType -> Maybe IQuery.RcImagesInfo -> (Maybe Text, Maybe Text)
-getImageMetaFromVehicleImage entityImagesInfo docType mbRcImagesInfo =
-  let images = case docType of
-        DVC.VehicleRegistrationCertificate ->
-          let merchantId = entityImagesInfo.merchantOperatingCity.merchantId
-           in IQuery.filterImagesByEntityAndType entityImagesInfo merchantId docType
-        _ -> case mbRcImagesInfo of
-          Just rcImagesInfo -> IQuery.filterRecentByPersonRCAndImageType rcImagesInfo docType
-          Nothing -> []
-      mbLatestImage = listToMaybe images
-   in (mbLatestImage <&> (.s3Path), mbLatestImage <&> (.id.getId))
 
 checkImageValidity :: IQuery.EntityImagesInfo -> DVC.DocumentType -> (Maybe ResponseStatus, Maybe Text, Maybe BaseUrl)
 checkImageValidity entityImagesInfo docType = do
@@ -1450,38 +1064,6 @@ checkLMSTrainingStatus driverId merchantOpCityId = do
   hasCompleted <- TPlasma.allLMSTrainingCompleted merchantOpCityId (driverId.getId)
   return $ hasCompleted >>= (\ok -> if ok then Just VALID else Nothing)
 
-checkVehiclePhotosStatusByRC :: Maybe IQuery.RcImagesInfo -> Flow (ResponseStatus, Maybe Text, Maybe BaseUrl)
-checkVehiclePhotosStatusByRC mbRcImagesInfo = do
-  -- Check all vehicle photo types from RC images (6 types: VehicleLeft, VehicleRight, VehicleFrontInterior, VehicleBackInterior, VehicleFront, VehicleBack)
-  let vehiclePhotoTypes = vehicleDocsByRcIdList
-      photoStatuses = case mbRcImagesInfo of
-        Just rcImagesInfo ->
-          mapMaybe
-            ( \docType -> do
-                let images = IQuery.filterRecentByPersonRCAndImageType rcImagesInfo docType
-                case images of
-                  [] -> Nothing -- No image for this type
-                  img : _ -> Just img.verificationStatus -- Get status of most recent image for this type
-            )
-            vehiclePhotoTypes
-        Nothing -> []
-  case photoStatuses of
-    [] -> return (NO_DOC_AVAILABLE, Nothing, Nothing)
-    statuses -> do
-      -- Check if all photos exist (must have all 6)
-      let allPhotosExist = length statuses == length vehiclePhotoTypes
-          allValid = allPhotosExist && all (== Just Documents.VALID) statuses
-          allManualVerification = allPhotosExist && all (== Just Documents.MANUAL_VERIFICATION_REQUIRED) statuses
-          -- Check if we have some VALID and all others are MANUAL_VERIFICATION_REQUIRED (no INVALID or missing)
-          hasValid = any (== Just Documents.VALID) statuses
-          onlyValidOrManual = all (\s -> s == Just Documents.VALID || s == Just Documents.MANUAL_VERIFICATION_REQUIRED) statuses
-          someValidAndRestManual = allPhotosExist && hasValid && onlyValidOrManual && not allValid
-      if allValid
-        then return (VALID, Nothing, Nothing)
-        else
-          if allManualVerification || someValidAndRestManual
-            then return (MANUAL_VERIFICATION_REQUIRED, Nothing, Nothing)
-            else return (INVALID, Nothing, Nothing)
 
 checkBackgroundVerificationStatus :: Id DP.Person -> Id DM.Merchant -> Id DMOC.MerchantOperatingCity -> Flow (ResponseStatus, Maybe Text, Maybe BaseUrl)
 checkBackgroundVerificationStatus driverId merchantId merchantOpCityId = do
@@ -1569,75 +1151,9 @@ getInProgressDriverDocuments driverId entityImagesInfo docType possibleVehicleCa
     _ -> return (NO_DOC_AVAILABLE, Nothing, Nothing)
   return (status, mbReason, mbUrl, Nothing, mbS3Path, mbImageId, Nothing)
 
-vehicleDocsByRcIdList :: [DVC.DocumentType]
-vehicleDocsByRcIdList =
-  [ DVC.VehicleLeft,
-    DVC.VehicleRight,
-    DVC.VehicleFrontInterior,
-    DVC.VehicleBackInterior,
-    DVC.VehicleFront,
-    DVC.VehicleBack,
-    DVC.Odometer
-  ]
 
-vehicleDocsLoadedByRcId :: [DVC.DocumentType]
-vehicleDocsLoadedByRcId =
-  vehicleDocsByRcIdList
-    <> [ DVC.VehiclePermit,
-         DVC.VehicleFitnessCertificate,
-         DVC.VehicleInsurance,
-         DVC.VehiclePUC,
-         DVC.VehicleNOC
-       ]
 
-getInProgressVehicleDocuments :: IQuery.EntityImagesInfo -> Maybe IQuery.RcImagesInfo -> DVC.DocumentType -> [DVC.DocumentVerificationConfig] -> Flow (ResponseStatus, Maybe Text, Maybe BaseUrl, Maybe UTCTime, Maybe Text, Maybe Text)
-getInProgressVehicleDocuments entityImagesInfo mbRcImagesInfo docType docVerificationConfigs = do
-  let onlyImageLookup = maybe False (fromMaybe False . (.onlyImageVerificationStatusLookupRequired)) $ find (\c -> c.documentType == docType) docVerificationConfigs
-      (mbS3Path, mbImageId) = case mbRcImagesInfo of
-        Just rcImagesInfo ->
-          let images = IQuery.filterRecentByPersonRCAndImageType rcImagesInfo docType
-              mbLatestImage = listToMaybe images
-           in (mbLatestImage <&> (.s3Path), mbLatestImage <&> (.id.getId))
-        Nothing -> getImageMetaFromLatestImage entityImagesInfo docType
-  (status, mbReason, mbUrl) <- case docType of
-    DVC.VehicleRegistrationCertificate -> checkIfUnderProgress entityImagesInfo DVC.VehicleRegistrationCertificate
-    DVC.SubscriptionPlan -> return (NO_DOC_AVAILABLE, Nothing, Nothing)
-    DVC.VehiclePermit -> return $ checkIfImageUploadedOrInvalidatedByRC mbRcImagesInfo DVC.VehiclePermit onlyImageLookup
-    DVC.VehicleFitnessCertificate -> return $ checkIfImageUploadedOrInvalidatedByRC mbRcImagesInfo DVC.VehicleFitnessCertificate onlyImageLookup
-    DVC.VehicleInsurance -> return $ checkIfImageUploadedOrInvalidatedByRC mbRcImagesInfo DVC.VehicleInsurance onlyImageLookup
-    DVC.VehiclePUC -> return $ checkIfImageUploadedOrInvalidatedByRC mbRcImagesInfo DVC.VehiclePUC onlyImageLookup
-    DVC.VehicleInspectionForm -> checkVehiclePhotosStatusByRC mbRcImagesInfo
-    DVC.VehicleNOC -> return $ checkIfImageUploadedOrInvalidatedByRC mbRcImagesInfo DVC.VehicleNOC onlyImageLookup
-    DVC.InspectionHub -> do
-      mbRegistrationNo <- case mbRcImagesInfo of
-        Just rcImagesInfo -> do
-          mbRc <- RCQuery.findById rcImagesInfo.rcId
-          case mbRc of
-            Just rc -> Just <$> decrypt rc.certificateNumber
-            Nothing -> return Nothing
-        Nothing -> return Nothing
-      case mbRegistrationNo of
-        Just registrationNo -> do
-          mbStatus <- getInspectionHubStatusForResponseStatus DOHR.ONBOARDING_INSPECTION Nothing (Just registrationNo)
-          let status = fromMaybe INVALID mbStatus
-          return (status, Nothing, Nothing)
-        Nothing -> return (NO_DOC_AVAILABLE, Nothing, Nothing)
-    _ | docType `elem` vehicleDocsByRcIdList -> return $ checkIfImageUploadedOrInvalidatedByRC mbRcImagesInfo docType onlyImageLookup
-    _ -> return (NO_DOC_AVAILABLE, Nothing, Nothing)
-  return (status, mbReason, mbUrl, Nothing, mbS3Path, mbImageId)
 
-checkIfImageUploadedOrInvalidatedByRC :: Maybe IQuery.RcImagesInfo -> DDVC.DocumentType -> Bool -> (ResponseStatus, Maybe Text, Maybe BaseUrl)
-checkIfImageUploadedOrInvalidatedByRC mbRcImagesInfo docType onlyImageLookup = do
-  let images = case mbRcImagesInfo of
-        Just rcImagesInfo -> IQuery.filterRecentByPersonRCAndImageType rcImagesInfo docType
-        _ -> []
-  case images of
-    [] -> (NO_DOC_AVAILABLE, Nothing, Nothing)
-    latestImage : _ -> do
-      case latestImage.verificationStatus of
-        Just Documents.VALID | onlyImageLookup -> (VALID, Nothing, Nothing)
-        Just Documents.INVALID -> (INVALID, extractImageFailReason latestImage.failureReason, Nothing)
-        _ -> (MANUAL_VERIFICATION_REQUIRED, Nothing, Nothing)
 
 checkIfImageUploadedOrInvalidated :: IQuery.EntityImagesInfo -> DDVC.DocumentType -> Bool -> DocVerificationConfigs -> Flow (ResponseStatus, Maybe Text, Maybe BaseUrl)
 checkIfImageUploadedOrInvalidated entityImagesInfo docType onlyImageLookup allDocVerificationConfigs = do
@@ -1666,77 +1182,8 @@ checkIfImageUploadedOrInvalidated entityImagesInfo docType onlyImageLookup allDo
             then return (FAILED, Nothing, Nothing)
             else return (MANUAL_VERIFICATION_REQUIRED, Nothing, Nothing)
 
-checkIfUnderProgress :: IQuery.EntityImagesInfo -> DVC.DocumentType -> Flow (ResponseStatus, Maybe Text, Maybe BaseUrl)
-checkIfUnderProgress entityImagesInfo docType = do
-  mbVerificationReqRecord <- case entityImagesInfo.entity of
-    IQuery.PersonEntity person -> do
-      -- Driver inspection or status handler: docType: DriverLicense/VehicleRegistrationCertificate
-      let driverId = person.id
-      mbVerificationReqIdfy <- listToMaybe <$> IVQuery.findLatestByDriverIdAndDocType (Just 1) Nothing driverId docType
-      mbVerificationReqHV <- listToMaybe <$> HVQuery.findLatestByDriverIdAndDocType (Just 1) Nothing driverId docType
-      pure $ getLatestVerificationRecord mbVerificationReqIdfy mbVerificationReqHV
-    IQuery.VehicleRCEntity _rc -> do
-      -- Vehicle inspection: docType: VehicleRegistrationCertificate
-      -- Inprogress rc doc for Vehicle inspection case not required, because we already have processed rc
-      pure Nothing
-  case mbVerificationReqRecord of
-    Just verificationReqRecord -> do
-      if verificationReqRecord.status == "pending" || verificationReqRecord.status == "source_down_retrying"
-        then return (PENDING, Nothing, Nothing)
-        else return (FAILED, verificationReqRecord.verificaitonResponse, Nothing)
-    Nothing -> do
-      let images = IQuery.filterRecentByEntityIdAndImageType entityImagesInfo docType
-      handleImages entityImagesInfo.transporterConfig.onboardingTryLimit images
-  where
-    handleImages onboardingTryLimit images
-      | null images = return (NO_DOC_AVAILABLE, Nothing, Nothing)
-      | length images > onboardingTryLimit * bool 1 2 (docType == DVC.DriverLicense || docType == DVC.AadhaarCard) = return (LIMIT_EXCEED, Nothing, Nothing)
-      | otherwise = do
-        let latestImage = head images
-        if latestImage.verificationStatus == Just Documents.INVALID
-          then return (INVALID, extractImageFailReason latestImage.failureReason, Nothing)
-          else return (NO_DOC_AVAILABLE, Nothing, Nothing)
 
-extractImageFailReason :: Maybe DriverOnboardingError -> Maybe Text
-extractImageFailReason imageError =
-  case imageError of
-    Just (ImageNotValid reason) -> Just reason -- only this because we are inserting this type only in manual reject request in update documents dashboard api
-    _ -> Nothing
 
-documentStatusMessage :: ResponseStatus -> Maybe Text -> DDVC.DocumentType -> Maybe BaseUrl -> Language -> Bool -> Flow (Maybe Text)
-documentStatusMessage status mbReason docType mbVerificationUrl language skipMessages
-  | skipMessages = pure Nothing
-  | otherwise = do
-    msg <- case (status, docType, mbVerificationUrl) of
-      (VALID, _, _) -> toVerificationMessage DocumentValid language
-      (MANUAL_VERIFICATION_REQUIRED, _, _) -> toVerificationMessage UnderManualReview language
-      (PENDING, DDVC.BackgroundVerification, Just _) -> toVerificationMessage VerificationPendingOnUserInput language
-      (PENDING, _, _) -> toVerificationMessage VerificationInProgress language
-      (PULL_REQUIRED, _, _) -> toVerificationMessage PullRequired language
-      (CONSENT_DENIED, _, _) -> toVerificationMessage ConsentDenied language
-      (LIMIT_EXCEED, _, _) -> toVerificationMessage LimitExceed language
-      (NO_DOC_AVAILABLE, _, _) -> toVerificationMessage NoDcoumentFound language
-      (INVALID, DDVC.DriverLicense, _) -> do
-        msg <- toVerificationMessage DLInvalid language
-        return $ fromMaybe msg mbReason
-      (INVALID, DDVC.VehicleRegistrationCertificate, _) -> do
-        msg <- toVerificationMessage RCInvalid language
-        return $ fromMaybe msg mbReason
-      (INVALID, _, _) -> do
-        msg <- toVerificationMessage DocumentInvalid language
-        return $ fromMaybe msg mbReason
-      (UNAUTHORIZED, _, _) -> toVerificationMessage Unauthorized language
-      (FAILED, _, _) -> do
-        case mbReason of
-          Just res
-            | "id_not_found" `T.isInfixOf` res -> toVerificationMessage InvalidDocumentNumber language
-            | "source_down" `T.isInfixOf` res -> toVerificationMessage VerificationInProgress language
-            | "TIMEOUT" `T.isInfixOf` res -> toVerificationMessage VerficationFailed language
-            | "BAD_REQUEST" `T.isInfixOf` res -> toVerificationMessage InvalidDocumentNumber language
-            | "422" `T.isInfixOf` res -> toVerificationMessage InvalidDocumentNumber language
-            | otherwise -> toVerificationMessage Other language
-          Nothing -> toVerificationMessage Other language
-    pure $ Just msg
 
 getAadhaarStatus :: Id DP.Person -> Flow (ResponseStatus, Maybe DAadhaarCard.AadhaarCard)
 getAadhaarStatus personId = do
@@ -1798,14 +1245,6 @@ getRCAndStatus driverId entityImagesInfo language = do
               msg <- toVerificationMessage NoDcoumentFound language
               return (NO_DOC_AVAILABLE, Nothing, msg)
 
-mapStatus :: Documents.VerificationStatus -> ResponseStatus
-mapStatus = \case
-  Documents.PENDING -> PENDING
-  Documents.MANUAL_VERIFICATION_REQUIRED -> MANUAL_VERIFICATION_REQUIRED
-  Documents.VALID -> VALID
-  Documents.INVALID -> INVALID
-  Documents.UNAUTHORIZED -> UNAUTHORIZED
-  Documents.PULL_REQUIRED -> PULL_REQUIRED
 
 verificationStatusCheck :: ResponseStatus -> Language -> DVC.DocumentType -> Maybe [Text] -> Flow Text
 verificationStatusCheck status language img mbReasons = do
@@ -1882,46 +1321,9 @@ getMessageFromResponse language response = do
       | otherwise -> toVerificationMessage Other language
     Nothing -> toVerificationMessage Other language
 
-data VerificationMessage
-  = InvalidDocumentNumber
-  | VerficationFailed
-  | NoDcoumentFound
-  | LimitExceed
-  | DLInvalid
-  | RCInvalid
-  | DocumentInvalid
-  | DocumentValid
-  | VerificationInProgress
-  | VerificationPendingOnUserInput
-  | UnderManualReview
-  | Unauthorized
-  | PullRequired
-  | ConsentDenied
-  | Other
-  | Reasons
-  deriving (Show, Eq, Ord)
 
-translateDynamicKey :: Text -> Language -> Flow Text
-translateDynamicKey key lang = do
-  mTranslation <- MTQuery.findByErrorAndLanguage key lang
-  return $ fromMaybe key (mTranslation <&> (.message))
 
-toVerificationMessage :: VerificationMessage -> Language -> Flow Text
-toVerificationMessage msg lang = do
-  errorTranslations <- MTQuery.findByErrorAndLanguage (T.pack (show msg)) lang
-  case errorTranslations of
-    Just errorTranslation -> return $ errorTranslation.message
-    Nothing -> return "Something went wrong"
 
-getLatestVerificationRecord :: Maybe IV.IdfyVerification -> Maybe HV.HyperVergeVerification -> Maybe SDO.VerificationReqRecord
-getLatestVerificationRecord mbIdfyVerificationReq mbHvVerificationReq = do
-  case (mbIdfyVerificationReq <&> (.createdAt), mbHvVerificationReq <&> (.createdAt)) of
-    (Just idfyCreatedAt, Just hvCreatedAt) -> if idfyCreatedAt > hvCreatedAt then SDO.makeIdfyVerificationReqRecord <$> mbIdfyVerificationReq else SDO.makeHVVerificationReqRecord <$> mbHvVerificationReq
-    (Nothing, Just _) -> SDO.makeHVVerificationReqRecord <$> mbHvVerificationReq
-    (Just _, Nothing) -> SDO.makeIdfyVerificationReqRecord <$> mbIdfyVerificationReq
-    (Nothing, Nothing) -> Nothing
-
--- Convert CommonDriverOnboardingDocuments to CommonDocumentItem
 mkCommonDocumentItem :: DCDOD.CommonDriverOnboardingDocuments -> CommonDocumentItem
 mkCommonDocumentItem doc =
   CommonDocumentItem
@@ -1958,26 +1360,5 @@ mapDigilockerToResponseStatus DocStatus.DOC_CONSENT_DENIED = Just CONSENT_DENIED
 mapDigilockerToResponseStatus DocStatus.DOC_PULL_REQUIRED = Just PULL_REQUIRED
 mapDigilockerToResponseStatus DocStatus.DOC_SUCCESS = Just VALID
 
-checkInspectionHubRequestCreated :: (EsqDBFlow m r, MonadFlow m, CacheFlow m r) => DOHR.RequestType -> Maybe (Id DP.Person) -> Maybe Text -> m (Maybe DOHR.RequestStatus)
-checkInspectionHubRequestCreated requestType mbDriverId mbRegistrationNo = do
-  mbInspectionReq <- case requestType of
-    DOHR.DRIVER_ONBOARDING_INSPECTION -> case mbDriverId of
-      Just driverId -> runInReplica $ QOHRE.findLatestByDriverIdAndRequestType driverId requestType
-      Nothing -> pure Nothing
-    DOHR.ONBOARDING_INSPECTION -> case mbRegistrationNo of
-      Just registrationNo -> runInReplica $ QOHRE.findLatestByRegistrationNoAndRequestType registrationNo requestType
-      Nothing -> pure Nothing
-    _ -> pure Nothing
-  pure $ (.requestStatus) <$> mbInspectionReq
 
-mapInspectionHubRequestStatusToResponseStatus :: Maybe DOHR.RequestStatus -> Maybe ResponseStatus
-mapInspectionHubRequestStatusToResponseStatus mbRequestStatus = case mbRequestStatus of
-  Just DOHR.APPROVED -> Just VALID
-  Just DOHR.PENDING -> Just VALID
-  Just DOHR.REJECTED -> Just INVALID
-  Nothing -> Just NO_DOC_AVAILABLE
 
-getInspectionHubStatusForResponseStatus :: (EsqDBFlow m r, MonadFlow m, CacheFlow m r) => DOHR.RequestType -> Maybe (Id DP.Person) -> Maybe Text -> m (Maybe ResponseStatus)
-getInspectionHubStatusForResponseStatus requestType mbDriverId mbRegistrationNo = do
-  mbRequestStatus <- checkInspectionHubRequestCreated requestType mbDriverId mbRegistrationNo
-  pure $ mapInspectionHubRequestStatusToResponseStatus mbRequestStatus
